DB/Creature: Warden Nork Bloodfrenzy & Son

+++ src/server/scripts/Custom/system_resurrection_dynamic.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2017 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "system_resurrection_dynamic.h"
+#include "../../../scripts/Custom/system_resurrection_dynamic.h"
 
 bool Dynamic_Resurrection::IsInDungeonOrRaid(Player* player)
 {
@@ -38,6 +38,6 @@ void Dynamic_Resurrection::DynamicResurrection(Player* player)
 	if (Creature* creature = player->FindNearestCreature(C_Resurrection_ENTRY, C_DISTANCE_CHECK_RANGE))
 		player->TeleportTo(player->GetMapId(), creature->GetPositionX(), creature->GetPositionY(), creature->GetPositionZ(), 1);
 	// Revive Player with 70 %
-	player->ResurrectPlayer( (float) 0.7);
+	player->ResurrectPlayer(0.7);
 	player->SpawnCorpseBones();
 }
